singular-bank-standalone-app/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
aiosqlite
kafka-python
kafka-python
